requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Flask>=2.0.0
Flask-CORS>=4.0.0
Pillow>=9.0.0
numpy>=1.21.0
Werkzeug>=2.0.0
gunicorn>=20.1.0
requests>=2.28.0